pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-nexus"
version = "1.1.0"
description = "Remote server management through MCP with built-in intelligence — the AI learns how you work."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Lightcap AI", email = "dev@lightcap.ai"},
]
keywords = ["mcp", "model-context-protocol", "remote-server", "ssh", "devops", "ai-tools", "claude", "infrastructure", "ai-memory", "intelligence"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "mcp>=1.0.0",
    "asyncssh>=2.14.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.30.0",
    "starlette>=0.37.0",
    "aiofiles>=24.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0",
    "ruff>=0.5.0",
    "mypy>=1.10",
]

[project.urls]
Homepage = "https://lightcap.ai"
Repository = "https://github.com/lightcap-ai/mcp-nexus"
Documentation = "https://github.com/lightcap-ai/mcp-nexus#readme"
Issues = "https://github.com/lightcap-ai/mcp-nexus/issues"

[project.scripts]
mcp-nexus = "mcp_nexus.__main__:main"

[tool.setuptools.packages.find]
include = ["mcp_nexus*"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
disallow_untyped_defs = false
